refactor: extract output.rs
